requirements.txt
requests>=2.31.0
openai>=1.30.0
pandas>=2.2.0
numpy>=1.26.0
scikit-learn>=1.4.0
scipy>=1.13.0
matplotlib>=3.8.0
seaborn>=0.13.0
sentence-transformers>=2.7.0
tqdm>=4.66.0
python-dotenv>=1.0.0
tabulate>=0.9.0
